@@ -1,44 +1,62 @@
-cmake_minimum_required(VERSION 3.9)
+cmake_minimum_required(VERSION 3.11)
 project(TopoLite)
 
 set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} ${CMAKE_CURRENT_SOURCE_DIR}/cmake)
-set(CMAKE_CXX_STANDARD 11)
-set(Boost_NO_BOOST_CMAKE 1)
 
-#if(LINUX)
-#    message(STATUS ">>> Linux")
-#    Set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -fPIC")
-#endif()
-#special flag for linux
+add_definitions(-DUNITTEST_DATAPATH="${CMAKE_CURRENT_SOURCE_DIR}/data")
+set(CMAKE_CXX_STANDARD 17)
 
 #########################################
 #####                               #####
-#####           Option              #####
+#####      Compiler flags           #####
 #####                               #####
 #########################################
 
-option(USE_PYTHON_PLUGIN "USE PYBIND11 PLUGIN" ON)
-option(USE_UNIT_TEST "USE CATCH2 TEST" ON)
-option(USE_C_EXTENSION "USE_C_EXTENSION" OFF)
+set(BUILD_TOPOGUI         ON CACHE STRING "Build gui" FORCE)
+set(BUILD_TOPOTEST        ON CACHE STRING "Build test" FORCE)
+set(BUILD_TOPOPYBIND      ON CACHE STRING "Build pybind11" FORCE)
 
-#########################################
-#####                               #####
-#####      External Library         #####
-#####                               #####
-#########################################
+if(NOT CMAKE_BUILD_TYPE)
+    set(CMAKE_BUILD_TYPE Release)
+endif()
 
+# set(CMAKE_CXX_FLAGS "-Wall -Wextra")
 
-#3rd party library
-if(USE_UNIT_TEST OR USE_PYTHON_PLUGIN OR USE_C_EXTENSION)
-    #For general purpose
-    set(TOPO_RUNTIME_OUTPUT ${CMAKE_CURRENT_SOURCE_DIR}/Release)
-    #For Rhino Plugin(Mac, Windows)
+# Handles GCC and CLANG - Note that GCC on macOS is actually an alias to LLVM CLANG
+
+if (${CMAKE_CXX_COMPILER_ID} MATCHES "Clang")
+    message("Compiling ${CMAKE_CXX_COMPILER_ID}")
+    set(CMAKE_CXX_FLAGS_DEBUG "-g -stdlib=libc++")
+    set(CMAKE_CXX_FLAGS_RELEASE "-O3 -stdlib=libc++")
+elseif(${CMAKE_CXX_COMPILER_ID} MATCHES "GNU")
+    message("Compiling with ${CMAKE_CXX_COMPILER_ID} GCC")
+    set(CMAKE_CXX_FLAGS_DEBUG "-g")
+    set(CMAKE_CXX_FLAGS_RELEASE "-O3")
+    # NANOGUI uses Metal, not supported by GCC
+    option(BUILD_TOPOGUI "Compile the gui" OFF)
+endif()
+
+    #3rd party library
+#if(USE_UNIT_TEST OR USE_PYTHON_PLUGIN OR USE_C_EXTENSION)
+#    set(TOPO_RUNTIME_OUTPUT ${CMAKE_CURRENT_SOURCE_DIR}/Release)
 #	if(MSVC)
 #		set(TOPO_RUNTIME_OUTPUT ${CMAKE_CURRENT_SOURCE_DIR}/../TopoRhino/WinBin)
 #	else()
 #		set(TOPO_RUNTIME_OUTPUT ${CMAKE_CURRENT_SOURCE_DIR}/../TopoRhino/MacBin)
 #	endif()
-endif()
+#endif()
+
+#########################################
+#####                               #####
+#####      Dependencies             #####
+#####                               #####
+#########################################
+
+# Download and define targets for third-party dependencies
+
+include(TopoliteDependencies)
+
+# Add deps directory
 
 add_subdirectory("${CMAKE_CURRENT_SOURCE_DIR}/ext/")
 
