@@ -15,17 +15,67 @@
 #ifndef _MESH_CREATOR_H
 #define _MESH_CREATOR_H
 
-#include "Utility/vec.h"
 #include <vector>
 #include "Mesh/CrossMesh.h"
 
-typedef shared_ptr<PolyMesh> pPolyMesh;
-typedef shared_ptr<_Polygon> pPolygon;
+enum PatternType{
+    CROSS_SQUARE = 1,
+    CROSS_RHOMBUS = 2,
+    CROSS_SQUARE_RHOMBUS = 3,
+    CROSS_HEXAGON = 4,
+    CROSS_HEXAGON_RHOMBUS = 5,
+    CROSS_DODECAGON_HEXAGON_QUAD = 6,
+    CROSS_OCTAGON_SQUARE = 7,
+    CROSS_OCTAGON_SQUARE_COLINEAR = 8,
+    CROSS_DODECAGON = 9,
+    CROSS_PENTAGON_CROSS = 10,
+    CROSS_PENTAGON_SNOW = 11,
+    CROSS_PENTAGON_MIRROR = 12
+};
 
-using namespace std;
-
+template<typename Scalar>
 class PatternCreator : public TopoObject
 {
+public:
+    typedef shared_ptr<PolyMesh<Scalar>> pPolyMesh;
+
+    typedef shared_ptr<_Polygon<Scalar>> pPolygon;
+
+    typedef shared_ptr<CrossMesh<Scalar>> pCrossMesh;
+
+    typedef Matrix<Scalar, 3, 1> Vector3;
+
+    typedef Matrix<Scalar, 2, 1> Vector2;
+    
+public:
+    
+    struct pVertex_compare
+    {
+        bool operator()(const Vector3 &A, const Vector3 &B) const
+        {
+            double eps = FLOAT_ERROR_LARGE;
+
+            if (A[0] - B[0] < -eps)
+                return true;
+            if (A[0] - B[0] > eps)
+                return false;
+
+            if (A[1] - B[1] < -eps)
+                return true;
+            if (A[1] - B[1] > eps)
+                return false;
+
+            if (A[2] - B[2] < -eps)
+                return true;
+            if (A[2] - B[2] > eps)
+                return false;
+
+            return false;
+        }
+    };
+
+    typedef std::set<Vector3, pVertex_compare> setVertex;
+
 public:
 	PatternCreator();
 	PatternCreator(shared_ptr<InputVarList> var):TopoObject(var){}
@@ -33,32 +83,25 @@ public:
 
 public:
 
-	// Create Mesh from Multiple Meshes
-	void CreateMesh_Merge(vector<pPolyMesh> polyMeshes, pPolyMesh &out);
-
 	// Create Mesh (2D Regular Pattern)
-	void CreateMesh_2DPattern(int patternID,
-	                          int patternRadius,
-	                          shared_ptr<CrossMesh> &out);
+	void create2DPattern(PatternType patternID,
+	                     int patternRadius,
+	                     pCrossMesh &out);
 
+    // Create Polygons
+    void createPolygonRoot(int edgeNum, Scalar edgeLen, pPolygon &out);
 
-public:
-	void CreateMesh_2DPattern(vector<_Polygon> &root_polys,
-                              Vector3f DX,
-                              Vector3f DY,
-                              int Nx,
-                              int Ny,
-                              shared_ptr<CrossMesh> &out);
+    // Compute Neighbors
+    void computeNeighbors(PatternType crosstype, pPolygon poly, vector<pPolygon> &out);
 
-	void CreateMesh_2DPattern_HexagonPattern2(int patternRadius, shared_ptr<CrossMesh> &out);
-    void CreateMesh_2DPattern_HexagonPattern3(int patternRadius, shared_ptr<CrossMesh> &out);
-    void CreateMesh_2DPattern_PentagonPattern3(int patternRadius, shared_ptr<CrossMesh> &out);
-	int GetPolygonIndexInList(pPolygon tagtPoly, vector<pPolygon> polyList);
+    bool checkPolygonExistance(pPolygon poly, const setVertex &vertices_set);
+    
+    void addToVerticesSet(pPolygon poly, setVertex &vertices_set);
+
+    Vector3 rotateVector(Vector3 rotCenter, Vector3 rotAxis, Scalar rotAngle, Vector3 tagtPt);
 
 public:
 
-	// Compute Neighbors
-	void ComputeNeighbors(int crosstype, pPolygon poly, vector<pPolygon> &out);
 	void ComputeNeighbor_Square(pPolygon poly, pPolygon &out);
 	void ComputeNeighbor_Hexagon(pPolygon poly, pPolygon &out);
 	void ComputeNeighbor_Octagon_Square(pPolygon poly, int edgeID, pPolygon &out);
@@ -70,32 +113,23 @@ public:
 	void ComputeNeighbor_Pentagon_Snow(pPolygon poly, int edgeID, pPolygon &out);
 	void ComputeNeighbor_Pentagon_Mirror(pPolygon poly, int edgeID, pPolygon &out);
 	void ComputeNeighbor_Rhombus(pPolygon poly, int edgeID, pPolygon &out);
-
     void ComputeNeighbor_Octagon_Square_Colinear(pPolygon poly, int edgeID, pPolygon &out);
 
-	Vector3f ComputeTileTranslation(pPolygon neighbor, Vector3f tagtStaPt, Vector3f tagtEndPt);
-    Vector3f ComputeTileTranslation_OCTAGON_SQUARE_COLINEAR(pPolygon poly, int edgeID);
+    Matrix<Scalar, 3, 1> ComputeTileTranslation(pPolygon neighbor, Vector3 tagtStaPt, Vector3 tagtEndPt);
+    Matrix<Scalar, 3, 1> ComputeTileTranslation_OCTAGON_SQUARE_COLINEAR(pPolygon poly, int edgeID);
 
-	// Create Polygons
-	void CreatePolygon_Root(int edgeNum, float edgeLen, pPolygon &out);
-	void CreatePolygon_Square(pPolygon &out, float edgeLen, int polyType = POLY_SQUARE_THETA_45);
-	void CreatePolygon_Hexagon(pPolygon &out, float edgeLen, int polyType = POLY_HEXAGON_TYPE_0);
-	void CreatePolygon_Octagon(pPolygon &out, float edgeLen, int polyType);
-	void CreatePolygon_Dodecagon(pPolygon &out, float edgeLen);
-	void CreatePolygon_Rhombus(pPolygon &out, float edgeLen, int polyType = POLY_RHOMBUS_THETA_0);
+	void CreatePolygon_Square(pPolygon &out, Scalar edgeLen, int polyType = POLY_SQUARE_THETA_45);
+	void CreatePolygon_Hexagon(pPolygon &out, Scalar edgeLen, int polyType = POLY_HEXAGON_TYPE_0);
+	void CreatePolygon_Octagon(pPolygon &out, Scalar edgeLen, int polyType);
+	void CreatePolygon_Dodecagon(pPolygon &out, Scalar edgeLen);
+	void CreatePolygon_Rhombus(pPolygon &out, Scalar edgeLen, int polyType = POLY_RHOMBUS_THETA_0);
 
-	void CreatePolygon_Pentagon_Cross(pPolygon &out, float edgeLen, int polyType = POLY_PENTAGON_CROSS_TYPE_0);
-	void CreatePolygon_Pentagon_Snow(pPolygon &out, float edgeLen, int polyType = POLY_PENTAGON_SNOW_TYPE_0);
-    void CreatePolygon_Pentagon_Mirror(pPolygon &out, float edgeLen, int polyType = POLY_PENTAGON_MIRROR_TYPE_0);
-
-	// Create Mesh (a Single Fan)
-	void CreateMesh_Fan(Vector3f edgeA, Vector3f edgeB, Vector3f edgeMid, pPolyMesh &out);
-	void CreatePolygon_Fan(Vector3f verA, Vector3f verB, Vector3f verM, pPolygon &out);
+	void CreatePolygon_Pentagon_Cross(pPolygon &out, Scalar edgeLen, int polyType = POLY_PENTAGON_CROSS_TYPE_0);
+	void CreatePolygon_Pentagon_Snow(pPolygon &out, Scalar edgeLen, int polyType = POLY_PENTAGON_SNOW_TYPE_0);
+    void CreatePolygon_Pentagon_Mirror(pPolygon &out, Scalar edgeLen, int polyType = POLY_PENTAGON_MIRROR_TYPE_0);
 
 public:
     //python
     vector<vector<double>> PyCreateMesh_2DPattern(int patternID, int patternRadius);
-
 };
-
 #endif
