@@ -2,217 +2,88 @@
 // Created by ziqwang on 14.01.19.
 //
 
-#include "ContactGraph.h"
 
-#include "Utility/PolyPolyBoolean.h"
-#include <iostream>
-#include <algorithm>
-#include <set>
-#include <tbb/tbb.h>
-#include <cmath>
+#include "ContactGraph.h"
+#include "Utility/ConvexHull2D.h"
+
 /*************************************************
 *
 *                  Basic Operation
 *
 *************************************************/
-using pPolyMesh = shared_ptr<PolyMesh>;
-ContactGraph::ContactGraph(shared_ptr<InputVarList> varList):TopoObject(varList)
-{
+
+/**
+ * @brief Class constructor
+ * @tparam Scalar
+ * @param varList
+ */
+template<typename Scalar>
+ContactGraph<Scalar>::ContactGraph(const shared_ptr<InputVarList> &varList):TopoObject(varList) {
+    
 }
 
-ContactGraph::~ContactGraph()
-{
+/**
+ * @brief Class destructor
+ * @tparam Scalar
+ */
+template<typename Scalar>
+ContactGraph<Scalar>::~ContactGraph() {
     nodes.clear();
     edges.clear();
+    dynamic_nodes.clear();
 }
 
-bool ContactGraph::constructFromPolyMeshes(vector<shared_ptr<PolyMesh>> &meshes,
+
+/**
+ * @brief build the graph from a list of polygonal meshes
+ * @tparam Scalar
+ * @param meshes
+ * @param atBoundary
+ * @param eps
+ * @param convexhull : merges faces which on a same plane by its convexhull
+ * @return
+ */
+template<typename Scalar>
+bool ContactGraph<Scalar>::buildFromMeshes(vector<pPolyMesh> &input_meshes,
                                            vector<bool> &atBoundary,
-                                           double eps)
-{
+                                           Scalar eps,
+                                           bool convexhull) {
 
-    //0) scale the meshes into united box
-    double maxD = 1;
-    for (int id = 0; id < meshes.size(); id++) {
-        pPolyMesh poly = meshes[id];
-        if (poly == nullptr) return false;
-        for (shared_ptr<_Polygon> face : poly->polyList) {
-            //1.1) construct plane
-            plane_contact plane;
-            Vector3f nrm = face->ComputeNormal();
-            Vector3f center = face->vers[0].pos;
-            plane.nrm = EigenPoint(nrm[0], nrm[1], nrm[2]);
+    // [1] - Scale meshes_input into a united box
+    meshes_input = input_meshes;
 
-            plane.D = nrm ^ center;
-            maxD = (std::max)(maxD, (double)std::abs(plane.D));
-        }
-    }
 
-    //1) create contact planes
-    vector<plane_contact> planes;
-    std::set<plane_contact, plane_contact_compare> setPlanes;
+    // [2] - cluster the faces of input mesh which are on a same plane
+    contact_faces.clear();
+    if(!clusterFacesofInputMeshes(eps))
+        return false;
 
-    int groupID = 0;
-    for (int id = 0; id < meshes.size(); id++)
-    {
-        pPolyMesh poly = meshes[id];
-        if (poly == nullptr)
-            return false;
-        for (shared_ptr<_Polygon> face : poly->polyList)
-        {
-            if(face->vers.size() < 3) continue;
-
-            //1.1) construct plane
-            plane_contact plane;
-            Vector3f nrm = face->ComputeNormal();
-            Vector3f center = face->vers[0].pos;
-            plane.nrm = EigenPoint(nrm[0], nrm[1], nrm[2]);
-            
-            plane.D = (nrm ^ center) / maxD;
-            plane.partID = id;
-            plane.polygon = face;
-            plane.eps = eps;
-
-            //1.2) find groupID
-            std::set<plane_contact, plane_contact_compare>::iterator find_it = setPlanes.end();
-            for(int reverse = -1; reverse <= 1; reverse += 2){
-                plane_contact tmp_plane = plane;
-                tmp_plane.nrm *= reverse;
-                tmp_plane.D *= reverse;
-                find_it = setPlanes.find(tmp_plane);
-                if(find_it != setPlanes.end()){
-                    plane.groupID = (*find_it).groupID;
-                    break;
-                }
-            }
-
-            if(find_it == setPlanes.end()){
-                plane.groupID = groupID ++;
-                setPlanes.insert(plane);
-            }
-            
-            planes.push_back(plane);
-        }
-    }
-
-    std::sort(planes.begin(), planes.end(), [&](const plane_contact &A, const plane_contact &B){
+    tbb::parallel_sort(contact_faces.begin(), contact_faces.end(), [&](const polygonal_face &A, const polygonal_face &B) {
         return A.groupID < B.groupID;
     });
 
-     //2) add nodes
-     for(int id = 0; id < meshes.size(); id++){
-         meshes[id]->ComputeVolume();
-         meshes[id]->ComputeCentroid();
-         EigenPoint centroid(meshes[id]->centroid.x, meshes[id]->centroid.y, meshes[id]->centroid.z);
-         pContactGraphNode node = make_shared<ContactGraphNode>(atBoundary[id], centroid, centroid, meshes[id]->volume);
-         addNode(node);
-     }
+    // [3] - find all pairs of potential face contacts
+    contact_pairs.clear();
+    listPotentialContacts(atBoundary);
 
-     //3) find all pairs of contact polygon
-     int sta = 0, end = 0;
-     vector<pairIJ> planeIJ;
-     while(sta < planes.size())
-     {
-          for(end = sta + 1; end < planes.size(); end++)
-          {
-              if(planes[sta].groupID != planes[end].groupID)
-              {
-                  break;
-              }
-          }
-          if (end - sta > 1)
-          {
-              for (int id = sta; id < end; id++)
-              {
-                  for(int jd = id + 1; jd < end; jd++)
-                  {
-                      int partI = planes[id].partID;
-                      int partJ = planes[jd].partID;
+    // [4] - compute the contacts
+    contact_graphedges.clear();
+    computeContacts();  // note: this function is vectorized with TBB
 
-                      EigenPoint nrmI = planes[id].nrm.normalized();
-                      EigenPoint nrmJ = planes[jd].nrm.normalized();
+    // [5] - add nodes into the graph
+    nodes.clear();
+    buildNodes(atBoundary);
 
-                      if(partI != partJ
-                      && std::abs(nrmI.dot(nrmJ) + 1) < FLOAT_ERROR_LARGE
-                      && (!atBoundary[partI] || !atBoundary[partJ]))
-                          planeIJ.push_back(pairIJ(id, jd));
-                  }
-              }
-          }
-          sta = end;
-      }
+    // [6] - add faces into graph
+    edges.clear();
+    buildEdges();
 
+    // [7] - assign node ID
+    finalize();
+    contact_edges = edges;
 
-     //4) parallel compute contacts
-     vector<shared_ptr<ContactGraphEdge>> planeIJEdges;
-     planeIJEdges.resize(planeIJ.size());
-
-     tbb::parallel_for(tbb::blocked_range<size_t>(0, planeIJ.size()), [&](const tbb::blocked_range<size_t>& r)
-	 {
-         for (size_t id = r.begin(); id != r.end(); ++id)
-         //for (size_t id = 0; id != planeIJ.size(); ++id)
-	 	{
-
-             int planeI = planeIJ[id].first;
-             int planeJ = planeIJ[id].second;
-
-             double projMat[16];
-             double invsProjMat[16];
-
-             pPolygon polyI = planes[planeI].polygon.lock();
-             pPolygon polyJ = planes[planeJ].polygon.lock();
-
-             polyI->ComputeProjectMatrixTo2D(projMat, invsProjMat);
-             vector<Vector3f> polyI2D = polyI->ProjectPolygonTo2D(projMat);
-             vector<Vector3f> polyJ2D = polyJ->ProjectPolygonTo2D(projMat);
-             vector<vector<Vector3f>> overlapPolyPts;
-
-             PolyPolyBoolean ppIntersec(getVarList());
-
-             ppIntersec.ComputePolygonsIntersection(polyI2D, polyJ2D, overlapPolyPts);
-             if (overlapPolyPts.size() == 0)
-                 continue;
-
-             vector<ContactPolygon> contactPolyEigens;
-             for(vector<Vector3f> overlapPoly: overlapPolyPts){
-                 vector<Vector3f> contactPoly;
-                 ppIntersec.ProjectPolygonTo3D(overlapPoly, invsProjMat, contactPoly);
-                 if(contactPoly.size() >= 3){
-                     ContactPolygon contactPolyEigen;
-                     EigenPoint ct(0, 0, 0);
-                     for(Vector3f pt: contactPoly)
-                     {
-                        EigenPoint ptEigen(pt[0], pt[1], pt[2]);
-                        contactPolyEigen.points.push_back(ptEigen);
-                        ct += ptEigen;
-                     }
-                     contactPolyEigen.center = ct/contactPoly.size();
-                     contactPolyEigens.push_back(contactPolyEigen);
-                 }
-             }
-
-             if(!contactPolyEigens.empty()){
-                 int partI = planes[planeI].partID;
-                 int partJ = planes[planeJ].partID;
-
-                 shared_ptr<ContactGraphEdge> edge = make_shared<ContactGraphEdge>(contactPolyEigens, planes[planeI].nrm);
-                 planeIJEdges[id] = edge;
-             }
-         }
-     });
-
-     //5) add contact edges
-     for(int id = 0; id < planeIJ.size(); id++)
-     {
-         pContactGraphEdge edge = planeIJEdges[id];
-         if(edge != nullptr){
-             int planeI = planeIJ[id].first;
-             int planeJ = planeIJ[id].second;
-             int partI = planes[planeI].partID;
-             int partJ = planes[planeJ].partID;
-             addContact(nodes[partI], nodes[partJ], edge);
-         }
-     }
+    // [8] - simplify contacts
+    if(convexhull) computeConvexHullofEdgePolygons();
 
     return true;
 }
@@ -222,18 +93,27 @@ bool ContactGraph::constructFromPolyMeshes(vector<shared_ptr<PolyMesh>> &meshes,
 *                  Graph Operation
 *
 *************************************************/
-
-void ContactGraph::addNode(shared_ptr<ContactGraphNode> _node)
-{
+/**
+ * @brief
+ * @tparam Scalar
+ * @param _node
+ */
+template<typename Scalar>
+void ContactGraph<Scalar>::addNode(pContactGraphNode _node) {
     _node->staticID = nodes.size();
 
     nodes.push_back(_node);
-
-    return;
 }
 
-void ContactGraph::addContact(shared_ptr<ContactGraphNode> _nodeA, shared_ptr<ContactGraphNode> _nodeB, shared_ptr<ContactGraphEdge> _edge)
-{
+/**
+ * @brief
+ * @tparam Scalar
+ * @param _nodeA
+ * @param _nodeB
+ * @param _edge
+ */
+template<typename Scalar>
+void ContactGraph<Scalar>::addContact(pContactGraphNode _nodeA, pContactGraphNode _nodeB, pContactGraphEdge _edge) {
 
     if (_nodeA->isBoundary && _nodeB->isBoundary)
         return;
@@ -252,48 +132,358 @@ void ContactGraph::addContact(shared_ptr<ContactGraphNode> _nodeA, shared_ptr<Co
     contactNeighbor.first = _nodeA;
     contactNeighbor.second = _edge;
     _nodeB->neighbors.push_back(contactNeighbor);
-
-    return;
 }
 
-void ContactGraph::finalize()
-{
+/**
+ * @brief merging nodeA and nodeB. they should move together.
+ * @tparam Scalar
+ * @param _nodeA,_nodeB
+ */
+template<typename Scalar>
+void ContactGraph<Scalar>::mergeNode(ContactGraph::pContactGraphNode _nodeA, ContactGraph::pContactGraphNode _nodeB) {
+    merged_nodes.push_back({_nodeA, _nodeB});
+}
+
+/**
+ * @brief Assign nodeID to the graph. Must be called after updating the graph
+ * @tparam Scalar
+ */
+template<typename Scalar>
+void ContactGraph<Scalar>::finalize() {
     int dynamicID = 0;
     dynamic_nodes.clear();
-    for (shared_ptr<ContactGraphNode> node : nodes)
-    {
-        if (!node->isBoundary)
-        {
+    for (pContactGraphNode node : nodes) {
+        if (!node->isBoundary) {
             node->dynamicID = dynamicID++;
             dynamic_nodes.push_back(node);
-        }
-        else
-        {
+        } else {
             node->dynamicID = -1;
         }
     }
 }
 
-void ContactGraph::getContactMesh(pPolyMesh &mesh)
-{
+
+
+
+/**
+ * @brief return the all contact polygons as a mesh
+ * @tparam Scalar
+ * @param mesh
+ */
+template<typename Scalar>
+void ContactGraph<Scalar>::getContactMesh(pPolyMesh &mesh) {
     mesh.reset();
-    mesh = make_shared<PolyMesh>(getVarList());
-    for(pContactGraphEdge edge: edges)
-    {
-        for(ContactPolygon poly: edge->polygons)
-        {
-            shared_ptr<_Polygon> face = make_shared<_Polygon>();
-            for(EigenPoint pt: poly.points)
-            {
-                face->push_back(Vector3f(pt[0], pt[1], pt[2]));
+    mesh = make_shared<PolyMesh<Scalar>>(getVarList());
+    for (pContactGraphEdge edge: contact_edges) {
+        for (pPolygon poly: edge->polygons) {
+            pPolygon copy_poly = make_shared<_Polygon<Scalar>>(*poly);
+            if(edge->partIDA > edge->partIDB){
+                copy_poly->reverseVertices();
             }
-            mesh->polyList.push_back(face);
+            mesh->polyList.push_back(copy_poly);
         }
     }
 
-    if(mesh)
-        mesh->removeDuplicatedVertices();
+    if (mesh) mesh->removeDuplicatedVertices();
 
 //    mesh->ScaleMesh(1.0 / normalized_scale);
 //    mesh->TranslateMesh(-normalized_trans);
 }
+
+
+
+
+/*************************************************
+*
+*             constructFromPolyMesh methods
+*
+*************************************************/
+
+/**
+ * @brief extract all polygonal faces from the meshes_input and cluster the faces which are on a same plane.
+ * @tparam Scalar
+ * @param eps
+ * @return
+ */
+template<typename Scalar>
+bool ContactGraph<Scalar>::clusterFacesofInputMeshes(Scalar eps)
+{
+    std::set<polygonal_face, plane_contact_compare> planes_set;
+
+    int groupID = 0;
+    size_t msize = meshes_input.size();
+    for (size_t id = 0; id < msize; id++)
+    {
+        pPolyMesh poly = meshes_input[id];
+        if (poly == nullptr)
+            return false;
+
+        for (pPolygon face : poly->polyList)
+        {
+            if (face->vers.size() < 3)
+                continue;
+
+            // 1.1) construct plane
+            polygonal_face plane;
+            Vector3 nrm = face->normal();
+            Vector3 center = face->vers[0]->pos;
+            plane.nrm = nrm;
+
+            plane.D = nrm.dot(center);
+            plane.partID = id;
+            plane.polygon = face;
+            plane.eps = eps;
+
+            // 1.2) find groupID
+            typename std::set<polygonal_face, plane_contact_compare>::iterator find_it = planes_set.end();
+            for (int reverse = -1; reverse <= 1; reverse += 2)
+            {
+                polygonal_face tmp_plane = plane;
+                tmp_plane.nrm *= reverse;
+                tmp_plane.D *= reverse;
+
+                find_it = planes_set.find(tmp_plane);
+                if (find_it != planes_set.end())
+                {
+                    plane.groupID = (*find_it).groupID;
+                    break;
+                }
+            }
+
+            if (find_it == planes_set.end())
+            {
+                plane.groupID = groupID++;
+                planes_set.insert(plane);
+            }
+
+            contact_faces.push_back(plane);
+        }
+    }
+    return true;
+}
+
+/**
+ * @brief find all pairs of faces which are on a same plane. Their normals have to be on opposite directions
+ * @tparam Scalar
+ * @param atBoundary
+ */
+template<typename Scalar>
+void ContactGraph<Scalar>::listPotentialContacts(vector<bool> &atBoundary)
+{
+    size_t sta = 0, end = 0;
+    while (sta < contact_faces.size())
+    {
+        for (end = sta + 1; end < contact_faces.size(); end++)
+        {
+            if (contact_faces[sta].groupID != contact_faces[end].groupID)
+            {
+                break;
+            }
+        }
+        if (end - sta > 1)
+        {
+            for (int id = sta; id < end; id++) {
+                for (int jd = id + 1; jd < end; jd++)
+                {
+                    int partI = contact_faces[id].partID;
+                    int partJ = contact_faces[jd].partID;
+
+                    Vector3 nrmI = contact_faces[id].nrm.normalized();
+                    Vector3 nrmJ = contact_faces[jd].nrm.normalized();
+
+                    if (partI != partJ
+                        && std::abs(nrmI.dot(nrmJ) + 1) < FLOAT_ERROR_LARGE
+                        && (!atBoundary[partI] || !atBoundary[partJ])){
+                        contact_pairs.push_back(partI < partJ ? pairIJ(id, jd) : pairIJ(jd, id));
+                    }
+
+                }
+            }
+        }
+        sta = end;
+    }
+}
+
+/**
+ * @brief Parallel compute contacts.
+ *        Though faces in the 'contact_pairs' are on the same plane,
+ *        checking whether they are overlap needs the 2D poly-poly intersection tool.
+ * @tparam Scalar
+ */
+template<typename Scalar>
+void ContactGraph<Scalar>::computeContacts()
+{
+    size_t psize = contact_pairs.size();
+    contact_graphedges.resize(psize);
+    //for (size_t id = 0; id < psize; ++id) // Sequential loop
+    tbb::parallel_for(tbb::blocked_range<size_t>(0, psize), [&](const tbb::blocked_range<size_t> &r)
+    {
+        for (size_t id = r.begin(); id != r.end(); ++id) {
+
+
+            int planeI = contact_pairs[id].first;
+            int planeJ = contact_pairs[id].second;
+
+            vector<Vector3> polyI = contact_faces[planeI].polygon.lock()->getVertices();
+            vector<Vector3> polyJ = contact_faces[planeJ].polygon.lock()->getVertices();
+
+            vector<vector<Vector3>> contactPtLists;
+
+            PolyPolyBoolean<Scalar> ppIntersec(getVarList());
+
+            ppIntersec.computePolygonsIntersection(polyI, polyJ, contactPtLists);
+            if (contactPtLists.empty())
+            {
+                continue;
+            }
+
+            vector<pPolygon> contactPolys;
+            for (vector<Vector3> contactPtList: contactPtLists)
+            {
+                if (contactPtList.size() >= 3) {
+                    pPolygon contactPoly = make_shared<_Polygon<Scalar>>();
+                    contactPoly->setVertices(contactPtList);
+                    contactPolys.push_back(contactPoly);
+                }
+            }
+
+            if (!contactPolys.empty())
+            {
+                contact_graphedges[id] = make_shared<ContactGraphEdge<Scalar>>(contactPolys, contact_faces[planeI].nrm);
+            }
+        }
+    });
+}
+
+/**
+ * @brief add each part as a node into the graph
+ * @tparam Scalar
+ * @param atBoundary
+ */
+template<typename Scalar>
+void ContactGraph<Scalar>::buildNodes(vector<bool> &atBoundary)
+{
+    for (size_t id = 0; id < meshes_input.size(); id++) {
+        Vector3 centroid = meshes_input[id]->centroid();
+        Scalar volume = meshes_input[id]->volume();
+        pContactGraphNode node = make_shared<ContactGraphNode<Scalar>>(atBoundary[id], centroid, centroid, volume);
+        addNode(node);
+    }
+}
+
+/**
+ * @brief for all pairs of potential contact faces,
+ *        if their contact_graphedges is not empty, then we add this edge into the graph
+ * @tparam Scalar
+ */
+template<typename Scalar>
+void ContactGraph<Scalar>::buildEdges() {
+
+    for (size_t id = 0; id < contact_pairs.size(); id++)
+    {
+        pContactGraphEdge edge = contact_graphedges[id];
+        if (edge != nullptr) {
+            int planeI = contact_pairs[id].first;
+            int planeJ = contact_pairs[id].second;
+            int partI = contact_faces[planeI].partID;
+            int partJ = contact_faces[planeJ].partID;
+            addContact(nodes[partI], nodes[partJ], edge);
+        }
+    }
+}
+
+/**
+ * @brief simplified the contact between two nodes, only compute their convex-hull
+ * @tparam Scalar
+ */
+template<typename Scalar>
+void ContactGraph<Scalar>::computeConvexHullofEdgePolygons()
+{
+    std::sort(contact_edges.begin(), contact_edges.end(), [&](pContactGraphEdge e0, pContactGraphEdge e1){
+        if(e0->partIDA < e1->partIDA)
+            return true;
+        if(e0->partIDA > e1->partIDA)
+            return false;
+
+        if(e0->partIDB < e1->partIDB)
+            return true;
+        if(e0->partIDB > e1->partIDB)
+            return false;
+
+        if (e0->normal[0] - e1->normal[0] < -FLOAT_ERROR_LARGE)
+            return true;
+        if (e0->normal[0] - e1->normal[0] > FLOAT_ERROR_LARGE)
+            return false;
+
+        if (e0->normal[1] - e1->normal[1] < -FLOAT_ERROR_LARGE)
+            return true;
+        if (e0->normal[1] - e1->normal[1] > FLOAT_ERROR_LARGE)
+            return false;
+
+        if (e0->normal[2] - e1->normal[2] < -FLOAT_ERROR_LARGE)
+            return true;
+        if (e0->normal[2] - e1->normal[2] > FLOAT_ERROR_LARGE)
+            return false;
+
+        return false;
+    });
+
+
+    edges.clear();
+    size_t sta = 0, end = 0;
+
+    while (sta < contact_edges.size()) {
+        for (end = sta + 1; end < contact_edges.size(); end++) {
+            if(!contact_edges[sta]->check_on_same_plane(contact_edges[end])){
+                break;
+            }
+        }
+
+        pContactGraphEdge edge;
+        if(end > sta + 1)
+        {
+            vector<Vector3> corner_points;
+            for(int id = sta; id < end; id++)
+            {
+                for(pPolygon poly: contact_edges[id]->polygons){
+                    vector<Vector3> poly_points = poly->getVertices();
+                    corner_points.insert(corner_points.end(), poly_points.begin(), poly_points.end());
+                }
+            }
+
+            Vector3 normal = contact_edges[sta]->normal, x_axis, y_axis;
+
+            vector<Vector3> convexhull_points;
+            ConvexHull2D<Scalar> convexhull_solver;
+            convexhull_solver.compute(corner_points, normal, convexhull_points);
+
+            pPolygon convexhull_poly = make_shared<_Polygon<Scalar>>();
+            convexhull_poly->setVertices(convexhull_points);
+
+            edge = make_shared<ContactGraphEdge<Scalar>>(convexhull_poly , contact_edges[sta]->normal);
+        }
+        else{
+            edge = make_shared<ContactGraphEdge<Scalar>>(contact_edges[sta]->polygons, contact_edges[sta]->normal);
+        }
+
+        edge->partIDA = contact_edges[sta]->partIDA;
+        edge->partIDB = contact_edges[sta]->partIDB;
+        edges.push_back(edge);
+        sta = end;
+    }
+}
+
+// No need to call this TemporaryFunction() function,
+// it's just to avoid link error.
+void TemporaryFunction_ContactGraph ()
+{
+    shared_ptr<InputVarList> varList;
+    ContactGraph<double> contactGraph(varList);
+    vector<shared_ptr<PolyMesh<double>>> polyMesh;
+    vector<bool> atBoundary;
+    contactGraph.buildFromMeshes(polyMesh, atBoundary);
+    contactGraph.mergeNode(nullptr, nullptr);
+    contactGraph.getContactMesh(polyMesh.front());
+}
+template class ContactGraph<double>;
+template class ContactGraph<float>;
