@@ -6,70 +6,60 @@
 #define TOPOLOCKCREATOR_CONTACTGRAPHEDGE_H
 
 #include <Eigen/Dense>
-#include <Eigen/StdVector>
 #include "Utility/HelpDefine.h"
-
-using EigenPoint =  Eigen::Vector3f;
-using EigenPointList = std::vector<Eigen::Vector3f,Eigen::aligned_allocator<Eigen::Vector3f>> ;
-using std::vector;
-
-class ContactPolygon{
-public:
-    EigenPointList points;
-    EigenPoint center;
-};
+#include <Mesh/Polygon.h>
 
 /*!
  * \brief: Stores which parts are contacted and the contact's geometry
  */
+template<typename Scalar>
 class ContactGraphEdge
 {
 public:
-    vector<ContactPolygon> polygons;  //!< a list which polygons are simple and closed (better to be convex)
-    vector<EigenPoint> normals; //!< contact normal from partIDA to partIDB
+
+    typedef shared_ptr<_Polygon<Scalar>> pPolygon;
+
+    typedef Matrix<Scalar, 3, 1> Vector3;
+
+public:
+    vector<pPolygon> polygons;  // a list which polygons are simple and closed (better to be convex)
+    Vector3 normal; // contact normal from partIDA to partIDB
 
 public:
 
-    ContactGraphEdge(ContactPolygon &_polygon, EigenPoint &_normal)
+    ContactGraphEdge(pPolygon &_polygon, Vector3 &_normal)
     {
         polygons.push_back(_polygon);
-        normals.push_back(_normal);
+        normal= _normal;
     }
 
-    ContactGraphEdge(vector<ContactPolygon> &_polygons, vector<EigenPoint> &_normals)
-    {
-        polygons = _polygons;
-        normals = _normals;
-    }
 
-    ContactGraphEdge(vector<ContactPolygon> &_polygons, EigenPoint &_normals)
+    ContactGraphEdge(vector<pPolygon> &_polygons, Vector3 &_normal)
     {
         polygons = _polygons;
-        for(int id = 0; id < polygons.size(); id++){
-            normals.push_back(_normals);
-        }
+        normal= _normal;
     }
 
     /*!
      * \return the contact normal starts from partID
      */
-    EigenPoint getContactNormal(int partID, int polyID){
+    Vector3 getContactNormal(int partID){
         if(partIDA == partID)
         {
-            return normals[polyID];
+            return normal;
         }
         else if(partIDB == partID){
-            return normals[polyID] * (-1.0f);
+            return normal * (-1.0f);
         }
-        return EigenPoint();
+        return Vector3(0, 0, 0);
     }
 
-    void get_norm_fric_for_block(int partID, int polyID, Eigen::Vector3d &normal, Eigen::Vector3d & ufric, Eigen::Vector3d & vfric)
+    void get_norm_fric_for_block(int partID, Vector3 &normal, Vector3 &ufric, Vector3 &vfric)
     {
-        normal = getContactNormal(partID, polyID).cast<double>() * (-1.0);
-        ufric = Eigen::Vector3d(1, 0, 0).cross(normal);
+        normal = getContactNormal(partID) * (-1.0);
+        ufric = Vector3(1, 0, 0).cross(normal);
         if((ufric).norm() < FLOAT_ERROR_SMALL){
-            ufric = Eigen::Vector3d(0, 1, 0).cross(normal);
+            ufric = Vector3(0, 1, 0).cross(normal);
         }
         ufric /= ufric.norm();
         vfric = normal.cross(ufric); vfric /= vfric.norm();
@@ -79,16 +69,34 @@ public:
     int num_points()
     {
         int num_vks = 0;
-        for(ContactPolygon poly : polygons)
-            num_vks += poly.points.size();
+        for(pPolygon poly : polygons)
+            num_vks += poly->size();
         return num_vks;
+    }
+
+    size_t size(){
+        return polygons.size();
+    }
+
+    bool check_on_same_plane(shared_ptr<ContactGraphEdge<Scalar>> e){
+        //always suppose partIDA < partIDB
+
+        if(partIDA != e->partIDA || partIDB != e->partIDB){
+            return false;
+        }
+
+        if(normal.cross(e->normal).norm() < FLOAT_ERROR_LARGE && normal.dot(e->normal) >= 0){
+            return true;
+        }
+
+        return false;
     }
 
 public: //Automatic Generate
 
-    int partIDA; //!< Static ID for start node of this edge
+    int partIDA; // Static ID for start node of this edge
 
-    int partIDB; //!< Static ID for end node of this edge
+    int partIDB; // Static ID for end node of this edge
 };
 
 
