@@ -10,104 +10,105 @@
 *
 *************************************************/
 
-void InterlockingSolver::get_force_from_norm_fric(Eigen::Vector3d n,       Eigen::Vector3d u,       Eigen::Vector3d v,
-                                            Eigen::RowVector2d &fkx, Eigen::RowVector2d &fky, Eigen::RowVector2d &fkz)
+template<typename Scalar>
+void InterlockingSolver<Scalar>::get_force_from_norm_fric(Vector3 n,       Vector3 u,       Vector3 v,
+                                            RowVector2 &fkx, RowVector2 &fky, RowVector2 &fkz)
 {
-    Eigen::Vector3d x_comp(1, 0 ,0);
-    Eigen::Vector3d y_comp(0, 1 ,0);
-    Eigen::Vector3d z_comp(0, 0 ,1);
-    fkx = Eigen::RowVector2d(n.dot(x_comp), -n.dot(x_comp));
-    fky = Eigen::RowVector2d(n.dot(y_comp), -n.dot(y_comp));
-    fkz = Eigen::RowVector2d(n.dot(z_comp), -n.dot(z_comp));
+    Vector3 x_comp(1, 0 ,0);
+    Vector3 y_comp(0, 1 ,0);
+    Vector3 z_comp(0, 0 ,1);
+    fkx = RowVector2(n.dot(x_comp), -n.dot(x_comp));
+    fky = RowVector2(n.dot(y_comp), -n.dot(y_comp));
+    fkz = RowVector2(n.dot(z_comp), -n.dot(z_comp));
     return;
 }
 
-void InterlockingSolver::get_force_from_norm_fric(Eigen::Vector3d    n,    Eigen::Vector3d   u,  Eigen::Vector3d   v,
-                                            Eigen::RowVector4d    &fkx, Eigen::RowVector4d &fky, Eigen::RowVector4d &fkz){
-    Eigen::Vector3d x_comp(1, 0 ,0);
-    Eigen::Vector3d y_comp(0, 1 ,0);
-    Eigen::Vector3d z_comp(0, 0 ,1);
-    fkx = Eigen::RowVector4d(n.dot(x_comp), -n.dot(x_comp), u.dot(x_comp), v.dot(x_comp));
-    fky = Eigen::RowVector4d(n.dot(y_comp), -n.dot(y_comp), u.dot(y_comp), v.dot(y_comp));
-    fkz = Eigen::RowVector4d(n.dot(z_comp), -n.dot(z_comp), u.dot(z_comp), v.dot(z_comp));
+template<typename Scalar>
+void InterlockingSolver<Scalar>::get_force_from_norm_fric(Vector3    n,    Vector3   u,  Vector3   v,
+                                            RowVector4    &fkx, RowVector4 &fky, RowVector4 &fkz){
+    Vector3 x_comp(1, 0 ,0);
+    Vector3 y_comp(0, 1 ,0);
+    Vector3 z_comp(0, 0 ,1);
+    fkx = RowVector4(n.dot(x_comp), -n.dot(x_comp), u.dot(x_comp), v.dot(x_comp));
+    fky = RowVector4(n.dot(y_comp), -n.dot(y_comp), u.dot(y_comp), v.dot(y_comp));
+    fkz = RowVector4(n.dot(z_comp), -n.dot(z_comp), u.dot(z_comp), v.dot(z_comp));
 }
 
-void InterlockingSolver::get_moment_from_norm_fric_vertex(Eigen::Vector3d n, Eigen::Vector3d u, Eigen::Vector3d v,
-                                                    Eigen::Vector3d r, Eigen::RowVector2d &mx, Eigen::RowVector2d &my,
-                                                    Eigen::RowVector2d &mz)
+template<typename Scalar>
+void InterlockingSolver<Scalar>::get_moment_from_norm_fric_vertex(Vector3 n, Vector3 u, Vector3 v,
+                                                    Vector3 r, RowVector2 &mx, RowVector2 &my,
+                                                    RowVector2 &mz)
 {
-    Eigen::Vector3d x_comp(1, 0 ,0);
-    Eigen::Vector3d y_comp(0, 1 ,0);
-    Eigen::Vector3d z_comp(0, 0 ,1);
-//    Eigen::Vector3d m_normal = n.cross(r);
-//    Eigen::Vector3d m_u_fric = u.cross(r);
-//    Eigen::Vector3d m_v_fric = v.cross(r);
-    Eigen::Vector3d m_normal = -r.cross(n);
-    Eigen::Vector3d m_u_fric = -r.cross(u);
-    Eigen::Vector3d m_v_fric = -r.cross(v);
+    Vector3 x_comp(1, 0 ,0);
+    Vector3 y_comp(0, 1 ,0);
+    Vector3 z_comp(0, 0 ,1);
+//    Vector3 m_normal = n.cross(r);
+//    Vector3 m_u_fric = u.cross(r);
+//    Vector3 m_v_fric = v.cross(r);
+    Vector3 m_normal = -r.cross(n);
+    Vector3 m_u_fric = -r.cross(u);
+    Vector3 m_v_fric = -r.cross(v);
 
-    mx = Eigen::RowVector2d(m_normal.dot(x_comp), -m_normal.dot(x_comp));
-    my = Eigen::RowVector2d(m_normal.dot(y_comp), -m_normal.dot(y_comp));
-    mz = Eigen::RowVector2d(m_normal.dot(z_comp), -m_normal.dot(z_comp));
+    mx = RowVector2(m_normal.dot(x_comp), -m_normal.dot(x_comp));
+    my = RowVector2(m_normal.dot(y_comp), -m_normal.dot(y_comp));
+    mz = RowVector2(m_normal.dot(z_comp), -m_normal.dot(z_comp));
 
     return;
 }
 
-void InterlockingSolver::get_moment_from_norm_fric_vertex(Eigen::Vector3d n,   Eigen::Vector3d u,   Eigen::Vector3d v, Eigen::Vector3d r,
-                                                    Eigen::RowVector4d &mx, Eigen::RowVector4d &my, Eigen::RowVector4d &mz)
+template<typename Scalar>
+void InterlockingSolver<Scalar>::get_moment_from_norm_fric_vertex(Vector3 n,   Vector3 u,   Vector3 v, Vector3 r,
+                                                    RowVector4 &mx, RowVector4 &my, RowVector4 &mz)
 {
 
-    Eigen::Vector3d x_comp(1, 0 ,0);
-    Eigen::Vector3d y_comp(0, 1 ,0);
-    Eigen::Vector3d z_comp(0, 0 ,1);
-    Eigen::Vector3d m_normal = -r.cross(n);
-    Eigen::Vector3d m_u_fric = -r.cross(u);
-    Eigen::Vector3d m_v_fric = -r.cross(v);
-    mx = Eigen::RowVector4d(m_normal.dot(x_comp), -m_normal.dot(x_comp), m_u_fric.dot(x_comp), m_v_fric.dot(x_comp));
-    my = Eigen::RowVector4d(m_normal.dot(y_comp), -m_normal.dot(y_comp), m_u_fric.dot(y_comp), m_v_fric.dot(y_comp));
-    mz = Eigen::RowVector4d(m_normal.dot(z_comp), -m_normal.dot(z_comp), m_u_fric.dot(z_comp), m_v_fric.dot(z_comp));
+    Vector3 x_comp(1, 0 ,0);
+    Vector3 y_comp(0, 1 ,0);
+    Vector3 z_comp(0, 0 ,1);
+    Vector3 m_normal = -r.cross(n);
+    Vector3 m_u_fric = -r.cross(u);
+    Vector3 m_v_fric = -r.cross(v);
+    mx = RowVector4(m_normal.dot(x_comp), -m_normal.dot(x_comp), m_u_fric.dot(x_comp), m_v_fric.dot(x_comp));
+    my = RowVector4(m_normal.dot(y_comp), -m_normal.dot(y_comp), m_u_fric.dot(y_comp), m_v_fric.dot(y_comp));
+    mz = RowVector4(m_normal.dot(z_comp), -m_normal.dot(z_comp), m_u_fric.dot(z_comp), m_v_fric.dot(z_comp));
 }
 
-void InterlockingSolver::get_A_j_k(int partID, int edgeID, Eigen::MatrixXd &Ajk, bool withFriction)
+template<typename Scalar>
+void InterlockingSolver<Scalar>::get_A_j_k(int partID, int edgeID, Eigen::MatrixXd &Ajk, bool withFriction)
 {
     if(!withFriction){
-        /*
-     * 1.0 get interface and part
-     */
-        shared_ptr<ContactGraphEdge> edge = graph->edges[edgeID];
-        shared_ptr<ContactGraphNode> part = graph->nodes[partID];
 
-        /*
-         * 2.0 initialize the matrix Ajk
-         */
+        // 1.0 get interface and part
+
+        shared_ptr<ContactGraphEdge<Scalar>> edge = graph->edges[edgeID];
+        shared_ptr<ContactGraphNode<Scalar>> part = graph->nodes[partID];
+
+        // 2.0 initialize the matrix Ajk
 
         int num_vks = edge->num_points();
         Ajk = Eigen::MatrixXd(6, 2 * num_vks);
 
         int stack_vk = 0;
-        for(int id = 0; id < edge->polygons.size(); id++)
+        for(size_t id = 0; id < edge->polygons.size(); id++)
         {
-            ContactPolygon poly = edge->polygons[id];
-            int num_vk = poly.points.size();
-            /*
-             * 3.0 filling the force fx, fy, fz, term
-             */
-            Eigen::Vector3d normal, u_fric, v_fric;
-            Eigen::RowVector2d fkx, fky, fkz;
-            edge->get_norm_fric_for_block(partID, id, normal, u_fric, v_fric);
-            get_force_from_norm_fric(normal, u_fric, v_fric, fkx, fky, fkz);
+            shared_ptr<_Polygon<Scalar>> poly = edge->polygons[id];
+            int num_vk = poly->vers.size();
+
+            // 3.0 filling the force fx, fy, fz, term
+
+            Matrix<Scalar, 3, 1> normal, u_fric, v_fric;
+            RowVector2 fkx, fky, fkz;
+            edge->get_norm_fric_for_block(partID, normal, u_fric, v_fric);
+            get_force_from_norm_fric(normal.template cast<double>(), u_fric.template cast<double>(), v_fric.template cast<double>(), fkx, fky, fkz);
             Ajk.block(0, 2 * stack_vk, 1, 2 * num_vk) = fkx.replicate(1, num_vk);
             Ajk.block(1, 2 * stack_vk, 1, 2 * num_vk) = fky.replicate(1, num_vk);
             Ajk.block(2, 2 * stack_vk, 1, 2 * num_vk) = fkz.replicate(1, num_vk);
 
-            /*
-             * 4.0 filling the torque mx, my, mz term
-             */
+            // 4.0 filling the torque mx, my, mz term
             for(int jd = 0; jd < num_vk; jd++)
             {
-                Eigen::Vector3d r = (poly.points[jd] - part->centroid).cast<double>();
-                Eigen::RowVector2d mx, my, mz;
-                get_moment_from_norm_fric_vertex(normal, u_fric, v_fric, r, mx, my, mz);
+                Vector3 r = (poly->vers[jd]->pos - part->centroid).template cast<double>();
+                RowVector2 mx, my, mz;
+                get_moment_from_norm_fric_vertex(normal.template cast<double>(), u_fric.template cast<double>(), v_fric.template cast<double>(), r, mx, my, mz);
                 Ajk.block(3, jd * 2 + 2 * stack_vk, 1, 2) = mx;
                 Ajk.block(4, jd * 2 + 2 * stack_vk, 1, 2) = my;
                 Ajk.block(5, jd * 2 + 2 * stack_vk, 1, 2) = mz;
@@ -115,43 +116,41 @@ void InterlockingSolver::get_A_j_k(int partID, int edgeID, Eigen::MatrixXd &Ajk,
             stack_vk += num_vk;
         }
     }
-    else{
-        /*
-        * 1.0 get interface and part
-        */
-        shared_ptr<ContactGraphEdge> edge = graph->edges[edgeID];
-        shared_ptr<ContactGraphNode> part = graph->nodes[partID];
+    else {
 
-        /*
-         * 2.0 initialize the matrix Ajk
-         */
+        // 1.0 get interface and part
+
+        shared_ptr<ContactGraphEdge<Scalar>> edge = graph->edges[edgeID];
+        shared_ptr<ContactGraphNode<Scalar>> part = graph->nodes[partID];
+
+        // 2.0 initialize the matrix Ajk
 
         int num_vks = edge->num_points();
         Ajk = Eigen::MatrixXd(6, 4 * num_vks);
         int stack_vk = 0;
-        for(int id = 0; id < edge->polygons.size(); id++)
+        for(size_t id = 0; id < edge->polygons.size(); id++)
         {
-            ContactPolygon poly = edge->polygons[id];
-            int num_vk = poly.points.size();
-            /*
-             * 3.0 filling the force fx, fy, fz, term
-             */
-            Eigen::Vector3d normal, u_fric, v_fric;
-            Eigen::RowVector4d fkx, fky, fkz;
-            edge->get_norm_fric_for_block(partID, id, normal, u_fric, v_fric);
-            get_force_from_norm_fric(normal, u_fric, v_fric, fkx, fky, fkz);
+            shared_ptr<_Polygon<Scalar>> poly = edge->polygons[id];
+            int num_vk = poly->vers.size();
+
+            // 3.0 filling the force fx, fy, fz, term
+
+            Matrix<Scalar, 3, 1> normal, u_fric, v_fric;
+            RowVector4 fkx, fky, fkz;
+            edge->get_norm_fric_for_block(partID, normal, u_fric, v_fric);
+            get_force_from_norm_fric(normal.template cast<double>(), u_fric.template cast<double>(), v_fric.template cast<double>(), fkx, fky, fkz);
             Ajk.block(0, 4 * stack_vk, 1, 4 * num_vk) = fkx.replicate(1, num_vk);
             Ajk.block(1, 4 * stack_vk, 1, 4 * num_vk) = fky.replicate(1, num_vk);
             Ajk.block(2, 4 * stack_vk, 1, 4 * num_vk) = fkz.replicate(1, num_vk);
 
-            /*
-             * 4.0 filling the torque mx, my, mz term
-             */
+
+            // 4.0 filling the torque mx, my, mz term
+
             for(int jd = 0; jd < num_vk; jd++)
             {
-                Eigen::Vector3d r = (poly.points[jd] - part->centroid).cast<double>();
-                Eigen::RowVector4d mx, my, mz;
-                get_moment_from_norm_fric_vertex(normal, u_fric, v_fric, r, mx, my, mz);
+                Vector3 r = (poly->vers[jd]->pos - part->centroid).template cast<double>();
+                RowVector4 mx, my, mz;
+                get_moment_from_norm_fric_vertex(normal.template cast<double>(), u_fric.template cast<double>(), v_fric.template cast<double>(), r, mx, my, mz);
                 Ajk.block(3, jd * 4 + 4 * stack_vk, 1, 4) = mx;
                 Ajk.block(4, jd * 4 + 4 * stack_vk, 1, 4) = my;
                 Ajk.block(5, jd * 4 + 4 * stack_vk, 1, 4) = mz;
@@ -159,7 +158,6 @@ void InterlockingSolver::get_A_j_k(int partID, int edgeID, Eigen::MatrixXd &Ajk,
             stack_vk += num_vk;
         }
     }
-
 
     return;
 }
@@ -171,17 +169,19 @@ void InterlockingSolver::get_A_j_k(int partID, int edgeID, Eigen::MatrixXd &Ajk,
 *
 *************************************************/
 
-void InterlockingSolver::computeTranslationalInterlockingMatrix(vector<EigenTriple> &tri, Eigen::Vector2i &size)
+template<typename Scalar>
+void InterlockingSolver<Scalar>::computeTranslationalInterlockingMatrix(vector<EigenTriple> &tri, Eigen::Vector2i &size)
 {
     int rowID = 0;
     tri.clear();
-    for(shared_ptr<ContactGraphEdge> edge: graph->edges) {
+    for(shared_ptr<ContactGraphEdge<Scalar>> edge: graph->edges) {
 
         int iA = graph->nodes[edge->partIDA]->dynamicID;
 
         int iB = graph->nodes[edge->partIDB]->dynamicID;
 
-        for (EigenPoint nrm : edge->normals) {
+        Vector3 nrm = (edge->normal).template cast<double>();
+        for (size_t id = 0; id < edge->size(); id++) {
             if (iA != -1) {
                 tri.push_back(EigenTriple(rowID, 3 * iA, -nrm[0]));
                 tri.push_back(EigenTriple(rowID, 3 * iA + 1, -nrm[1]));
@@ -199,30 +199,32 @@ void InterlockingSolver::computeTranslationalInterlockingMatrix(vector<EigenTrip
     size = Eigen::Vector2i(rowID, 3 * graph->dynamic_nodes.size());
 }
 
-void InterlockingSolver::computeRotationalInterlockingMatrix(vector<EigenTriple> &tri, Eigen::Vector2i &size)
+template<typename Scalar>
+void InterlockingSolver<Scalar>::computeRotationalInterlockingMatrix(vector<EigenTriple> &tri, Eigen::Vector2i &size)
 {
     int rowID = 0;
     tri.clear();
-    for(shared_ptr<ContactGraphEdge> edge: graph->edges)
+    for(shared_ptr<ContactGraphEdge<Scalar>> edge: graph->edges)
     {
-        shared_ptr<ContactGraphNode> nodeA = graph->nodes[edge->partIDA];
-        shared_ptr<ContactGraphNode> nodeB = graph->nodes[edge->partIDB];
+        shared_ptr<ContactGraphNode<Scalar>> nodeA = graph->nodes[edge->partIDA];
+        shared_ptr<ContactGraphNode<Scalar>> nodeB = graph->nodes[edge->partIDB];
 
         int iA = nodeA->dynamicID;
         int iB = nodeB->dynamicID;
 
-        EigenPoint ctA = nodeA->centroid;
-        EigenPoint ctB = nodeB->centroid;
+        Vector3 ctA = (nodeA->centroid).template cast<double>();
+        Vector3 ctB = (nodeB->centroid).template cast<double>();
 
-        for(int id = 0; id < edge->normals.size(); id++)
+        Vector3 nrm = (edge->normal).template cast<double>();
+        for(size_t id = 0; id < edge->size(); id++)
         {
-            ContactPolygon poly = edge->polygons[id];
-            EigenPoint nrm = edge->normals[id];
-            for(EigenPoint pt: poly.points)
+            shared_ptr<_Polygon<Scalar>> poly = edge->polygons[id];
+            for(pVertex ver: poly->vers)
             {
+                Vector3 pt = (ver->pos).template cast<double>();;
                 if (iA != -1)
                 {
-                    EigenPoint mt = (pt - ctA).cross(nrm);
+                    Vector3 mt = (pt - ctA).cross(nrm);
                     tri.push_back(EigenTriple(rowID, 6 * iA, -nrm[0]));
                     tri.push_back(EigenTriple(rowID, 6 * iA + 1, -nrm[1]));
                     tri.push_back(EigenTriple(rowID, 6 * iA + 2, -nrm[2]));
@@ -232,7 +234,7 @@ void InterlockingSolver::computeRotationalInterlockingMatrix(vector<EigenTriple>
                 }
 
                 if (iB != -1) {
-                    EigenPoint mt = (pt - ctB).cross(nrm);
+                    Vector3 mt = (pt - ctB).cross(nrm);
                     tri.push_back(EigenTriple(rowID, 6 * iB, nrm[0]));
                     tri.push_back(EigenTriple(rowID, 6 * iB + 1, nrm[1]));
                     tri.push_back(EigenTriple(rowID, 6 * iB + 2, nrm[2]));
@@ -247,7 +249,8 @@ void InterlockingSolver::computeRotationalInterlockingMatrix(vector<EigenTriple>
     size = Eigen::Vector2i(rowID, 6 * graph->dynamic_nodes.size());
 }
 
-void InterlockingSolver::computeTranslationalInterlockingMatrixDense(Eigen::MatrixXd &mat){
+template<typename Scalar>
+void InterlockingSolver<Scalar>::computeTranslationalInterlockingMatrixDense(Eigen::MatrixXd &mat){
     vector<EigenTriple> tris;
     Eigen::Vector2i size;
 
@@ -262,7 +265,8 @@ void InterlockingSolver::computeTranslationalInterlockingMatrixDense(Eigen::Matr
     return;
 }
 
-void InterlockingSolver::computeRotationalInterlockingMatrixDense(Eigen::MatrixXd &mat)
+template<typename Scalar>
+void InterlockingSolver<Scalar>::computeRotationalInterlockingMatrixDense(Eigen::MatrixXd &mat)
 {
     vector<EigenTriple> tris;
     Eigen::Vector2i size;
@@ -279,7 +283,8 @@ void InterlockingSolver::computeRotationalInterlockingMatrixDense(Eigen::MatrixX
 }
 
 
-void InterlockingSolver::computeRotationalInterlockingMatrixSparse(EigenSpMat &spatMat)
+template<typename Scalar>
+void InterlockingSolver<Scalar>::computeRotationalInterlockingMatrixSparse(EigenSpMat &spatMat)
 {
     vector<EigenTriple> tris;
     Eigen::Vector2i size;
@@ -290,30 +295,82 @@ void InterlockingSolver::computeRotationalInterlockingMatrixSparse(EigenSpMat &s
 }
 
 
+template<typename Scalar>
+void InterlockingSolver<Scalar>::appendAuxiliaryVariables(vector<EigenTriple> &tri, Eigen::Vector2i &size)
+{
+    int num_row = size[0];
+    int num_col = size[1];
 
-void InterlockingSolver::computeEquilibriumMatrix(Eigen::MatrixXd &Aeq,  bool withFriction) {
+    for(size_t id = 0; id < num_row; id++){
+        tri.push_back(EigenTriple(id, num_col + id, -1));
+    }
+    size[1] += num_row;
+    return;
+}
+
+template<typename Scalar>
+void InterlockingSolver<Scalar>::appendMergeConstraints(vector<EigenTriple> &tri, Eigen::Vector2i &size, bool isRotation)
+{
+    int dimension = (isRotation ? 6 : 3);
+    for(int id = 0;id < graph->merged_nodes.size(); id++)
+    {
+        wpContactGraphNode A = graph->merged_nodes[id].first;
+        wpContactGraphNode B = graph->merged_nodes[id].second;
+        int iA = A.lock()->dynamicID;
+        int iB = B.lock()->dynamicID;
+        if(iA != -1 && iB != -1)
+        {
+            for(int index = 0; index < dimension; index++)
+            {
+                tri.push_back(EigenTriple(size[0] + id * dimension * 2 + index, dimension * iA + index, 1));
+                tri.push_back(EigenTriple(size[0] + id * dimension * 2 + index, dimension * iB + index, -1));
+
+                tri.push_back(EigenTriple(size[0] + id * dimension * 2 + index + 1, dimension * iA + index, -1));
+                tri.push_back(EigenTriple(size[0] + id * dimension * 2 + index + 1, dimension * iB + index, 1));
+            }
+        }
+        else{
+            for(int index = 0; index < dimension; index++){
+                if(iA != -1) {
+                    tri.push_back(EigenTriple(size[0] + id * dimension * 2 + index, dimension * iA + index, 1));
+                    tri.push_back(EigenTriple(size[0] + id * dimension * 2 + index + 1, dimension * iA + index, -1));
+                }
+                if(iB != -1){
+                    tri.push_back(EigenTriple(size[0] + id * dimension + index, dimension * iB + index, 1));
+                    tri.push_back(EigenTriple(size[0] + id * dimension * 2 + index + 1, dimension * iB + index, -1));
+
+                }
+            }
+        }
+    }
+    size[0] += graph->merged_nodes.size() * dimension * 2;
+    return;
+}
+
+template<typename Scalar>
+void InterlockingSolver<Scalar>::computeEquilibriumMatrix(Eigen::MatrixXd &Aeq,  bool withFriction) {
     if (!withFriction)
     {
-        /*
-         * 1.0 init the matrix
-         */
+
+        // 1.0 init the matrix
+
         vector<int> row_start_index;
         int start_index = 0;
-        for (shared_ptr<ContactGraphEdge> edge: graph->edges) {
+        for (shared_ptr<ContactGraphEdge<Scalar>> edge: graph->edges) {
             row_start_index.push_back(start_index);
-            //std::cout << edge->partIDA << ",\t" << edge->partIDB << ",\t" << edge->num_points() << std::endl;
+            // std::cout << edge->partIDA << ",\t" << edge->partIDB << ",\t" << edge->num_points() << std::endl;
             start_index += edge->num_points() * 2;
         }
         Aeq = Eigen::MatrixXd::Zero(graph->dynamic_nodes.size() * 6, start_index);
 
-        /*
-         * 2.0 build matrix
-         */
 
-        for (int id = 0; id < graph->edges.size(); id++) {
-            shared_ptr<ContactGraphEdge> edge = graph->edges[id];
-            shared_ptr<ContactGraphNode> partA = graph->nodes[edge->partIDA];
-            shared_ptr<ContactGraphNode> partB = graph->nodes[edge->partIDB];
+        // 2.0 build matrix
+
+
+        for (size_t id = 0; id < graph->edges.size(); id++) {
+            shared_ptr<ContactGraphEdge<Scalar>> edge = graph->edges[id];
+            shared_ptr<ContactGraphNode<Scalar>> partA = graph->nodes[edge->partIDA];
+            shared_ptr<ContactGraphNode<Scalar>> partB = graph->nodes[edge->partIDB];
             int dyn_partIDA = partA->dynamicID;
             int dyn_partIDB = partB->dynamicID;
             int num_vk = edge->num_points();
@@ -331,26 +388,24 @@ void InterlockingSolver::computeEquilibriumMatrix(Eigen::MatrixXd &Aeq,  bool wi
     }
 
     else{
-        /*
-         * 1.0 init the matrix
-         */
+
+        // 1.0 init the matrix
+
         vector<int> row_start_index;
         int start_index = 0;
-        for (shared_ptr<ContactGraphEdge> edge: graph->edges) {
+        for (shared_ptr<ContactGraphEdge<Scalar>> edge: graph->edges) {
             row_start_index.push_back(start_index);
-            //std::cout << edge->partIDA << ",\t" << edge->partIDB << ",\t" << edge->num_points() << std::endl;
+            // std::cout << edge->partIDA << ",\t" << edge->partIDB << ",\t" << edge->num_points() << std::endl;
             start_index += edge->num_points() * 4;
         }
         Aeq = Eigen::MatrixXd::Zero(graph->dynamic_nodes.size() * 6, start_index);
 
-        /*
-         * 2.0 build matrix
-         */
+        // 2.0 build matrix
 
-        for (int id = 0; id < graph->edges.size(); id++) {
-            shared_ptr<ContactGraphEdge> edge = graph->edges[id];
-            shared_ptr<ContactGraphNode> partA = graph->nodes[edge->partIDA];
-            shared_ptr<ContactGraphNode> partB = graph->nodes[edge->partIDB];
+        for (size_t id = 0; id < graph->edges.size(); id++) {
+            shared_ptr<ContactGraphEdge<Scalar>> edge = graph->edges[id];
+            shared_ptr<ContactGraphNode<Scalar>> partA = graph->nodes[edge->partIDA];
+            shared_ptr<ContactGraphNode<Scalar>> partB = graph->nodes[edge->partIDB];
             int dyn_partIDA = partA->dynamicID;
             int dyn_partIDB = partB->dynamicID;
             int num_vk = edge->num_points();
@@ -369,7 +424,8 @@ void InterlockingSolver::computeEquilibriumMatrix(Eigen::MatrixXd &Aeq,  bool wi
 
 }
 
-double InterlockingSolver::computeEquilibriumMatrixConditonalNumber()
+template <typename Scalar>
+Scalar InterlockingSolver<Scalar>::computeEquilibriumMatrixConditonalNumber()
 {
     Eigen::MatrixXd Aeq;
     computeEquilibriumMatrix(Aeq);
@@ -390,3 +446,19 @@ double InterlockingSolver::computeEquilibriumMatrixConditonalNumber()
 
 }
 
+
+// No need to call this TemporaryFunction() function,
+// it's just to avoid link error.
+void TemporaryFunction_InterlockingSolver ()
+{
+    InterlockingSolver<double> solver(nullptr, nullptr);
+    vector<InterlockingSolver<double>::EigenTriple> tri;
+    Eigen::Vector2i size;
+    bool isRotation = true;
+    solver.appendMergeConstraints(tri, size, isRotation);
+    solver.appendAuxiliaryVariables(tri, size);
+    solver.computeRotationalInterlockingMatrix(tri, size);
+    solver.computeTranslationalInterlockingMatrix(tri, size);
+}
+template class InterlockingSolver<double>;
+template class InterlockingSolver<float>;
