@@ -17,59 +17,32 @@
 
 
 #include <vector>
-#include "TopoLite/Structure/Part.h"
-#include "TopoLite/Utility/vec.h"
-#include "TopoLite/Utility/HelpStruct.h"
-#include "TopoLite/Utility/TopoObject.h"
-#include "TopoLite/Structure/Struc.h"
-#include "TopoLite/CrossMesh/CrossMeshCreator.h"
+#include "ConvexBlock.h"
+#include "Mesh/CrossMesh.h"
 
 using namespace std;
 
-using pCrossMeshCreator = shared_ptr<CrossMeshCreator> ;
-using pStruc = shared_ptr<Struc> ;
-
+template<typename Scalar>
 class StrucCreator : public TopoObject
 {
+public:
+    typedef shared_ptr<CrossMesh<Scalar>> pCrossMesh;
+    typedef shared_ptr<ConvexBlock<Scalar>> pConvexBlock;
+    typedef shared_ptr<Cross<Scalar>> pCross;
+    typedef Matrix<Scalar, 2, 1> Vector2;
+
 
 public:
-
-    pCrossMeshCreator   crossMeshCreator;             // Reference surface model
-	pStruc              struc;                        // Designed interlocking structure
-	double              worldAxesMat[16];
-	double              worldMatrix[16];
+    vector<pConvexBlock> blocks;
 
 public:
 
     StrucCreator(shared_ptr<InputVarList> var);
     ~StrucCreator();
-	StrucCreator(const StrucCreator &_myStruc);
-	void ClearStruc();
 
 public:
     // Create Structure
-
-	bool LoadSurface(const char* objFileName);
-    int CreateStructure(bool createCrossMesh,
-                        double interactMatrix[],
-                        bool previewMode);
-    int UpdateStructureGeometry(bool previewMode);
-
-public:
-    // Compute Part Contacts
-
-    void BuildPartsGraph();
-	void ComputeBoundaryParts();
-	void ComputePartContacts();
-	void ComputePartFaceContacts();
-    void ComputePartFaceContactsBruteForce();
-	void ComputePartEdgeContacts();
-
-public:
-    // Save Output Models
-
-	void WriteStructure(const char *folderPath);
-    void WriteCrossMesh(char *meshFileName);
+    bool compute(pCrossMesh crossMesh);
 };
 
 #endif
