@@ -16,7 +16,9 @@
 #define _HELP_DEFINE_H
 
 #include <math.h>
-
+#ifndef UNITTEST_DATAPATH
+    #define UNITTEST_DATAPATH
+#endif
 ///////////////////////////////////////////////////////////////////////
 //
 //  Debug Flag
@@ -27,6 +29,9 @@
 
 ///////////////////////////////////////////////////////////////////////
 //  General Definition
+
+// List
+#define ELEMENT_OUT_LIST            -1
 
 // File and buffer
 #define FILE_NAME_LENGTH            1024
@@ -118,49 +123,6 @@
 //#define CROSS_RING_NUM           40
 #define CROSS_L                  0.4
 
-
-//#define CROSS_TRI                0
-#define CROSS_SQUARE                       1
-#define CROSS_RHOMBUS                      2
-#define CROSS_SQUARE_RHOMBUS               3
-#define CROSS_HEXAGON                      4
-#define CROSS_HEXAGON_RHOMBUS              5
-#define CROSS_DODECAGON_HEXAGON_QUAD       6
-#define CROSS_OCTAGON_SQUARE               7
-
-#define CROSS_OCTAGON_SQUARE_COLINEAR      8
-#define CROSS_DODECAGON                    9
-#define CROSS_PENTAGON_CROSS              10
-#define CROSS_PENTAGON_SNOW               11
-#define CROSS_PENTAGON_MIRROR             12
-
-#define POLY_SQUARE_THETA_45               1
-#define POLY_SQUARE_THETA_15               2
-#define POLY_SQUARE_THETA_75               3
-#define POLY_RHOMBUS_THETA_0               4
-#define POLY_RHOMBUS_THETA_90              5
-#define POLY_RHOMBUS_THETA_120             6
-#define POLY_RHOMBUS_THETA_240             7
-#define POLY_HEXAGON_TYPE_0                8
-#define POLY_HEXAGON_TYPE_1                9
-#define POLY_OCTAGON_REGULAR              10
-#define POLY_OCTAGON_COLINEAR             11
-#define POLY_DODECAGON                    12
-#define POLY_PENTAGON_CROSS_TYPE_0        13
-#define POLY_PENTAGON_CROSS_TYPE_1        14
-#define POLY_PENTAGON_CROSS_TYPE_2        15
-#define POLY_PENTAGON_CROSS_TYPE_3        16
-#define POLY_PENTAGON_SNOW_TYPE_0         17
-#define POLY_PENTAGON_SNOW_TYPE_1         18
-#define POLY_PENTAGON_SNOW_TYPE_2         19
-#define POLY_PENTAGON_SNOW_TYPE_3         20
-#define POLY_PENTAGON_SNOW_TYPE_4         21
-#define POLY_PENTAGON_SNOW_TYPE_5         22
-#define POLY_PENTAGON_MIRROR_TYPE_0       23
-#define POLY_PENTAGON_MIRROR_TYPE_1       24
-#define POLY_RHOMBUS_TYPE_0               25
-#define POLY_RHOMBUS_TYPE_1               26
-
 #define MOBILITY_SCORE_MAX       3.0
 #define MOBILITY_SCORE_DIFF      0.0001
 
@@ -169,6 +131,7 @@
 #define NONE_GROUP              -1
 #define NONE_ELEMENT            -1
 #define NONE_FACE               -1
+#define NOT_FOUND               -1
 
 #define MOBILITY_SPHERE_R        3.0f
 
